+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Docstrings and doctestable examples to `record.py`.
+- Inputs can be validated using operations
+- `validate` parameter in `Input` takes `Operation.instance_name`
 - Logistic Regression with SAG optimizer
+
 ### Fixed
 - New model tutorial mentions file paths that should be edited.